backend/requirements.txt
aiosqlite==0.19.0
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.2.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
ecdsa==0.19.0
email-validator==2.0.0.post2
fastapi==0.103.2
greenlet==3.1.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
iniconfig==2.0.0
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
psycopg2==2.9.9
pyasn1==0.6.1
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pytest==7.4.2
pytest-asyncio==0.21.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.21
starlette==0.27.0
typing_extensions==4.12.2
uvicorn==0.30.6
